backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
redis>=5.2.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
httpx>=0.28.0
sse-starlette>=2.1.0
structlog>=24.4.0
prometheus-fastapi-instrumentator>=7.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
aiosqlite>=0.20.0
